changed to f32

+++ src/models/party.rs
@@ -45,32 +45,32 @@ impl PartyResultInt {
 
 #[derive(Debug)]
 pub struct PartyResultFloat {
-    pub spd: f64,
-    pub cdu: f64,
-    pub afd: f64,
-    pub fdp: f64,
-    pub gruene: f64,
-    pub linke: f64,
-    pub wg1: f64,
-    pub wg2: f64,
-    pub wg3: f64,
-    pub wg4: f64,
-    pub wg_sum: f64,
+    pub spd: f32,
+    pub cdu: f32,
+    pub afd: f32,
+    pub fdp: f32,
+    pub gruene: f32,
+    pub linke: f32,
+    pub wg1: f32,
+    pub wg2: f32,
+    pub wg3: f32,
+    pub wg4: f32,
+    pub wg_sum: f32,
 }
 
 impl PartyResultFloat {
     pub fn new(
-        spd: f64,
-        cdu: f64,
-        afd: f64,
-        fdp: f64,
-        gruene: f64,
-        linke: f64,
-        wg1: f64,
-        wg2: f64,
-        wg3: f64,
-        wg4: f64,
-        wg_sum: f64,
+        spd: f32,
+        cdu: f32,
+        afd: f32,
+        fdp: f32,
+        gruene: f32,
+        linke: f32,
+        wg1: f32,
+        wg2: f32,
+        wg3: f32,
+        wg4: f32,
+        wg_sum: f32,
     ) -> PartyResultFloat {
         PartyResultFloat {
             spd,
@@ -87,3 +87,4 @@ impl PartyResultFloat {
         }
     }
 }
+
